+++ example/jar/HelloWorld/Main.java
@@ -1,0 +1,7 @@
+package HelloWorld;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world");
+    }
+}