+++ src/helperFunctions.ts
@@ -1,13 +1,13 @@
-const os = require('os');
+const os = require("os");
 
 export function getExecutable(): string {
-    const platform = os.platform();
-    switch (platform) {
-        case 'darwin':
-            return 'Contents/Home/bin/java';
-        case 'win32': 
-            return 'bin/java.exe';
-        default:
-            return 'bin/java';  // for example linux
-      }
+  const platform = os.platform();
+  switch (platform) {
+    case "darwin":
+      return "Contents/Home/bin/java";
+    case "win32":
+      return "bin/java.exe";
+    default:
+      return "bin/java"; // for example linux
+  }
 }